I should be copying tensors for map

+++ src/WhiskerToolbox/ML_Widget/mlpack_conversion.hpp
@@ -156,9 +156,10 @@ inline arma::Mat<double> convertTensorDataToMlpackMatrix(
     // Fill the matrix with the tensor data
     for (std::size_t i = 0; i < numRows; ++i) {
         auto tensor = tensor_data.getTensorAtTime(timestamps[i]);
-        auto flattened_tensor = tensor.flatten().to(torch::kDouble);
+        //auto flattened_tensor = tensor.flatten().to(torch::kDouble);
+        auto flattened_tensor = tensor.flatten();
         for (std::size_t j = 0; j < flattened_tensor.numel(); ++j) {
-            result(i, j) = flattened_tensor[j].item<double>();
+            result(i, j) = flattened_tensor[j].item<float>();
         }
     }
